apps/functions/requirements.txt
firebase-functions>=0.1.0
firebase-admin>=6.5.0
flask>=3.0.0
python-jose[cryptography]>=3.3.0
openai>=1.30.1
python-dotenv>=1.0.1
